Migrate settings menu code to its own module
